requirements.txt
gevent
pyjwt
jsonschema
pymongo
pyyaml